Fix HSL to RGB conversion when hue < 60 (#1)
# Problem
When converting from HSL to RGB, the RGB values do not always come out as expected. This is only an issue when the original color's hue value is less than 60, which, when divided by `60.0` in the `remainder` function, results in a value less than 1 but greater than or equal to 0.

# Solution
Leverage Erlang's `:math.fmod` function in the `remainder` function to properly handle cases where dividing the hue by `60.0` results in a non-whole number.